Added path_env and load_conda scripts
The install.sh script now creates empty path_env and load_conda scripts
that are meant to contain the PATH (and other env variables) exports and
the conda init script respectively. In the future, it would be useful to
automate the conda init step to populate the load_conda.sh script.